Added delete and block buttons for users list

+++ src/main/java/com/epam/hotel/services/implementations/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package com.epam.hotel.services.implementations;
 
 import com.epam.hotel.domains.User;
+import com.epam.hotel.domains.enums.BlockStatus;
 import com.epam.hotel.domains.enums.Permission;
 import com.epam.hotel.repositories.UserRepository;
 import com.epam.hotel.services.UserService;
@@ -47,7 +48,12 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteById(Long id) {
-        userRepository.deleteById(id);
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.get().getPermission() == Permission.USER){
+            userRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("You cant delete ADMIN");
+        }
     }
 
     @Override
@@ -57,6 +63,20 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public Page<User> getAllUsersPaged(int page, int size) { return userRepository.findAll(PageRequest.of(page - 1, size, Sort.Direction.ASC, "id")); }
+
+    @Override
+    public void blockById(Long id) {
+        User user = userRepository.getOne(id);
+        user.setBlock(BlockStatus.BLOCKED);
+        userRepository.save(user);
+    }
+
+    @Override
+    public void unblockById(Long id) {
+        User user = userRepository.getOne(id);
+        user.setBlock(BlockStatus.UNBLOCKED);
+        userRepository.save(user);
+    }
 
 }
